Move setup_editor test fixture to editor module.
