+++ spyder/plugins/editor/widgets/tests/test_editor.py
@@ -19,7 +19,7 @@ import pytest
 from qtpy.QtCore import Qt
 
 # Local imports
-from spyder.utils.fixtures import setup_editor
+from spyder.plugins.editor.widgets.tests.fixtures import setup_editor
 from spyder.plugins.editor.widgets.editor import EditorStack
 from spyder.widgets.findreplace import FindReplace
