+++ spyder/utils/fixtures.py
@@ -19,9 +19,6 @@ except ImportError:
 # Third party imports
 import pytest
 
-# Local imports
-from spyder.plugins.editor.widgets.editor import EditorStack
-from spyder.widgets.findreplace import FindReplace
 from spyder.config.user import UserConfig
 from spyder.config.main import CONF_VERSION, DEFAULTS
 
@@ -47,22 +44,3 @@ def tmpconfig(request):
 
     request.addfinalizer(fin)
     return CONF
-
-@pytest.fixture
-def setup_editor(qtbot):
-    """
-    Set up EditorStack with CodeEditor containing some Python code.
-    The cursor is at the empty line below the code.
-    Returns tuple with EditorStack and CodeEditor.
-    """
-    text = ('a = 1\n'
-            'print(a)\n'
-            '\n'
-            'x = 2')  # a newline is added at end
-    editorStack = EditorStack(None, [])
-    editorStack.set_introspector(Mock())
-    editorStack.set_find_widget(FindReplace(editorStack))
-    editorStack.set_io_actions(Mock(), Mock(), Mock(), Mock())
-    finfo = editorStack.new('foo.py', 'utf-8', text)
-    qtbot.addWidget(editorStack)
-    return editorStack, finfo.editor
